requirements.txt
pandas
numpy
python3-bs4
shutil